+++ SpaceInvaders/missile.h
@@ -10,10 +10,10 @@ public:
 	void Update();
 	void SetState(bool state) { this->m_state = state; }
 	bool GetState() { return this->m_state; }
-	uint8_t GetDamage() { return this->m_damage; }
+	int GetDamage() { return this->m_damage; }
 		
 private:
 	bool m_state;
-	uint8_t m_damage = 1;
+	int m_damage = 1;
 };
